MODIFY type

+++ src/main/java/com/portfoliomaker/controller/PortfolioController.java
@@ -22,7 +22,7 @@ public class PortfolioController {
     }
 
     @PostMapping()
-    public String addPortfolio(String name, double price) {
+    public String addPortfolio(String name, long price) {
         portfolioService.save(new Portfolio(name, price));
         return "redirect:";
     }
